% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_residual_interactions.R
\name{add_residual_interactions}
\alias{add_residual_interactions}
\title{add_residual_interactions}
\usage{
add_residual_interactions(formula, data)
}
\arguments{
\item{formula}{a regression formula including the interaction terms}

\item{data}{a data.frame used for the regression}
}
\value{
Object containing the updated regression formula \code{obj$formula}
and data set \code{obj$data}
}
\description{
Creates residual interaction terms and adds them to the data.frame
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
