% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi.lm.R
\name{bootmi.lm}
\alias{bootmi.lm}
\alias{bootmi.lm.default}
\title{calculates regression analysis for bootstrapped and imputed data}
\usage{
bootmi.lm(bootmi)

\method{bootmi}{lm.default}(bootmi)
}
\arguments{
\item{bootmi}{An object of class \code{\link{bootmi}}}
}
\value{
object of class "bootmi.lm" including
\itemize{
  \item First Regression formula
  \item Second Original data
  \item Third Bootstrapped coefficients
  \item Fourth Bootstrap samples 
  \item Fifth Amount of bootstrap replications
  \item Sixth Moderation centered, TRUE or FALSE
  \item Seventh Seed value
}
}
\description{
The method regresses the model on original and bootstrapped data.
}
\details{
For faster computation, \pkg{parallel} is implemented.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
