% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi_ci.R
\name{bootmi_ci}
\alias{bootmi_ci}
\title{Helper function to use of boot.ci from boot package.}
\usage{
bootmi_ci(boot_object)
}
\arguments{
\item{boot_object}{object of class bootmi}
}
\value{
list with information for boot-package.
}
\description{
Helper function to use of boot.ci from boot package.
}
