% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi_results.R
\name{bootmi_results}
\alias{bootmi_results}
\title{creates output table for bootmi.lm}
\usage{
bootmi_results(bootmi_object)
}
\arguments{
\item{bootmi_object}{object of class "simslop.bootmi" or 
"bootmi.lm"}
}
\description{
This is a helper function of \code{\link{summary.bootmi.lm}} and 
\code{\link{summary.simslop.bootmi}} that creates output 
of the analyses.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
