% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_residual.R
\name{build_residual}
\alias{build_residual}
\title{creates residuals}
\usage{
build_residual(variable, data, new_varname)
}
\arguments{
\item{variable}{an interaction term}

\item{data}{a data.frame}

\item{new_varname}{name of the residual interaction term}
}
\value{
A data.frame of one row consisting of the residual interaction term.
}
\description{
Helper function of \code{\link{add_residual_interactions}} to 
create a specific residual interaction term.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
