% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bootmi.R
\name{calc_bootmi}
\alias{calc_bootmi}
\title{Prepare calculation of bootstrapped confidence intervals.}
\usage{
calc_bootmi(data, indices, frmla, imputationMethod, res_int = FALSE,
  center_mods = FALSE, simslopinfo = NA)
}
\arguments{
\item{boot_object}{object of class "simpleslopes.bootmi" or 
"bootmi.lm"}
}
\description{
This is a helper function to make use of \code{\link[boot]{boot.ci}} method 
from boot package.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
