% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.R
\name{centering}
\alias{centering}
\title{centers selected variables}
\usage{
centering(data_set, centered_vars)
}
\arguments{
\item{data_set}{a data.frame}

\item{centered_vars}{a vector containing names of variables to center}
}
\value{
data.frame with centered variables
}
\description{
It is important to center data that is used in the model 
AFTER deleting cases of missing values.
}
\references{
\href{http://processmacro.org/faq.html}{Process}
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
