% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.bootmi.R
\name{glm.bootmi}
\alias{glm.bootmi}
\title{Generalized linear model for bootstrapped and imputed data}
\usage{
glm.bootmi(bootmi, family = gaussian, ...)
}
\arguments{
\item{bootmi}{An object of class \code{\link{bootmi}}}

\item{family}{see \code{\link[stats]{glm}}}

\item{...}{any default glm statement}
}
\value{
object of class "bootmi.lm" including
\itemize{
  \item Regression formula
  \item Original data
  \item Bootstrapped coefficients
  \item Bootstrap samples 
  \item Amount of bootstrap replications
  \item Moderation centered, TRUE or FALSE
  \item Seed value
}
}
\description{
The method applies \code{glm()} on original and bootstrapped data.
}
\details{
For faster computation, \pkg{parallel} is implemented.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
