% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numformat.R
\name{numformat}
\alias{numformat}
\title{Round to decimal points}
\usage{
numformat(val, decpoints = 2)
}
\arguments{
\item{val}{numeric value}

\item{decpoints}{Round to how many decimal points}
}
\value{
\code{NULL}
}
\description{
Function to round to x decimal points including trailing zero
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
