% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootmi.R
\name{print.bootmi}
\alias{print.bootmi}
\alias{print.simpleslopes.bootmi}
\alias{print.bootmi.lm}
\alias{print.regosi}
\alias{print.simpleslopes}
\title{Prints a \code{bootmi} object}
\usage{
\method{print}{bootmi}(bootmi, ...)

\method{print}{simpleslopes.bootmi}(bootmi.slopes, ...)

\method{print}{bootmi.lm}(bootmi.lm, ...)

\method{print}{regosi}(regosi, ...)

\method{print}{simpleslopes}(slopes, ...)
}
\arguments{
\item{object}{A \code{bootmi} object}
}
\value{
\code{NULL}
}
\description{
Prints a \code{bootmi} object

Prints a \code{simpleslopes.bootmi} object

Prints a \code{bootmi.lm} object

Prints a \code{regosi} object

Prints a \code{simpleslopes} object
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
