% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regosi.R
\name{regosi}
\alias{regosi}
\alias{regosi.default}
\alias{regosi.lmerMod}
\alias{regosi.lm}
\alias{regosi.mira}
\alias{regosi.bootmi.lm}
\title{Calculates regions of significance for moderation effects}
\usage{
regosi(object, x_var, m_var, ci = 95)

\method{regosi}{default}(object, x_var, m_var, ci = 95)

\method{regosi}{lmerMod}(object, x_var, m_var, ci = 95)

\method{regosi}{lm}(object, x_var, m_var, ci = 95)

\method{regosi}{mira}(object, x_var, m_var, ci = 95)

\method{regosi}{bootmi.lm}(object, x_var, m_var, ci = 95)
}
\arguments{
\item{object}{Object containing $coeff, $dat, $cov_matrix}

\item{x_var}{Name of the independend variable}

\item{m_var}{Name of the moderating variable}

\item{ci}{Confidence interval, default 95}
}
\value{
object of class "regosi" including
\itemize{
  \item $a, Indicator for Interval
  \item $b, Value of slope
  \item $root_term, Indicator for significant regions
  \item $m_var, Name of the moderating variable
  \item $x_var, Name of the independend variable
}
}
\description{
The method is an implementation of regions of significance  
as proposed by Bauer and Curran (2005). The default takes 
parameters needed to calculate regions of significance. For objects 
of class "lm", "lmerMod" and "mira" exist helper functions 
that extract parameters "coeff", "dat", "cov_matrix" from the object.
}
\details{
CAUTION: 
For calculation of regions of significance, the ACOV-matrix
is used, instead of bootstrapped confidence intervals.
Hence, this method omits bootstraps from bootmi.lm and calculates 
regions of significance for imputed orginal model only. 
I am searching for a solution to the problem. 
Any suggestions are welcome:
\href{https://github.com/svolpers/bootmi/issues}{Suggest Solution}
For more reliable regions, use \code{\link{simslop}} with dicrete
values of the moderating variable, which reflect the whole range 
of the variable.
}
\section{Helper functions exist}{

Uses object of classes lm, lmerMod or mira 
and extracts coefficients, data, and variance covariance matrix
}

\references{
Bauer, Daniel J.; Curran, Patrick J. (2005): 
Probing Interactions in Fixed and Multilevel Regression: 
Inferential and Graphical Techniques. In: Multivariate Behavioral Research
40 (3), S. 373-400.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
