% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resimpcen.R
\name{resimpcen}
\alias{resimpcen}
\title{adds residual interactions, imputes and centers}
\usage{
resimpcen(frmla, data, res_int = TRUE, imputation, center_mods = FALSE,
  bootstraps = FALSE)
}
\arguments{
\item{frmla}{A regression formula}

\item{data}{A data.frame used for the regression}

\item{res_int}{TRUE or FALSE. Calculate residual interactions?}

\item{imputation}{Imputation method provided by mice-package.}

\item{center_mods}{TRUE or FALSE. Center interaction terms?}

\item{bootstraps}{TRUE or FALSE. Bootstrap sample or not?}
}
\value{
If bootstraps = FALSE: Returns list of data (list$data) and 
updated formula (list$formula)
If bootstraps = TRUE: Returns data
}
\description{
This is a helper function of \code{\link{bootmi}} that
\enumerate{
  \item calculates the (residual) interaction terms,
  \item imputes missing values using mice, and 
  \item centers interaction terms
}
}
\references{
Hippel, Paul T. von (2009): How to Impute Interactions,
Squares, and other Transformed Variables. In: 
Sociological Methodology 39 (1), S. 265-291.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
