% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleslopes.R
\name{simpleslopes}
\alias{simpleslopes}
\alias{simpleslopes.default}
\alias{simpleslopes.lm}
\alias{simpleslopes.mira}
\alias{simpleslopes.lmerMod}
\alias{simpleslopes.bootmi.lm}
\title{Conducts analysis of the simple slopes}
\usage{
simpleslopes(coeff, dat, cov_matrix, x_var, m_var, y_var, ci = 95,
  mod_values_type = c("sd", "val"), mod_values = c(-1, 0, 1),
  centered = FALSE)

\method{simpleslopes}{default}(coeff, dat, cov_matrix, x_var, m_var, y_var,
  ci = 95, mod_values_type = c("sd", "val"), mod_values = c(-1, 0, 1),
  centered = FALSE)

\method{simpleslopes}{lm}(object, x_var, m_var, ci = 95,
  mod_values_type = c("sd", "val"), mod_values = c(-1, 0, 1),
  centered = FALSE)

\method{simpleslopes}{mira}(object, x_var, m_var, ci = 95,
  mod_values_type = c("sd", "val"), mod_values = c(-1, 0, 1),
  centered = FALSE)

\method{simpleslopes}{lmerMod}(object, x_var, m_var, ci = 95,
  mod_values_type = c("sd", "val"), mod_values = c(-1, 0, 1),
  centered = FALSE)

\method{simpleslopes}{bootmi.lm}(bootmi_lm_obj, x_var, m_var, ci = 95,
  mod_values_type = "sd", mod_values = c(-1, 0, 1), centered = TRUE)
}
\arguments{
\item{coeff}{Coefficients from fitted model, just for default}

\item{dat}{Data used to fit model, just for default}

\item{cov_matrix}{Variance-Covariance-Matrix of fitted model, 
just for default}

\item{x_var}{Name of the independend variable}

\item{m_var}{Name of the moderating variable}

\item{y_var}{Name of the dependend variable}

\item{ci}{Confidence interval, default 95}

\item{mod_values_type}{Either sd (=standard deviation, default) or 
val (=values of data used)}

\item{mod_values}{Vector of values of the moderator, default c(-1,0,1)}

\item{centered}{Interaction coefficent mean centered? TRUE or FALSE}

\item{object}{Object, if availiable but NOT for default}
}
\value{
object of class "simpleslopes"
}
\description{
The method is an implementation of analysis of the simple slopes  
as proposed by Preacher, Curran, and Bauer (2006). The default takes 
parameters needed to conduct analysis of the simple slopes. For objects 
of class "bootmi.lm", "lm", "lmerMod" and "mira" exist helper functions 
that extract parameters "coeff", "dat", "cov_matrix" from the object.
}
\section{Helper functions exist}{
 Uses object to extract coefficients, 
data, dependend variable and variance covariance matrix
}

\references{
Preacher, Kristopher J.; Curran, Patrick J.; Bauer, Daniel J. 
(2006): Computational Tools for Probing Interactions in Multiple Linear 
Regression, Multilevel Modeling, and Latent Curve Analysis. In: Journal of 
Educational and Behavioral Statistics 31 (4), S. 437–448. 
DOI: 10.3102/10769986031004437.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
