% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootmi.R
\name{summary.bootmi}
\alias{summary.bootmi}
\alias{summary.simslop.bootmi}
\title{Summary of a \code{bootmi.lm} object}
\usage{
\method{summary}{bootmi}(object, ...)

\method{summary}{simslop.bootmi}(object, ci_type = c("basic", "norm",
  "stud", "perc", "bca", "all"), ci = 95, ...)
}
\arguments{
\item{object}{A \code{bootmi} object}

\item{...}{Other summary arguments}
}
\value{
\code{NULL}
}
\description{
Summary of a \code{bootmi} object

Summary of a \code{simslop.bootmi} object
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
