% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regosi.R
\name{summary.regosi}
\alias{summary.regosi}
\title{Summary of a \code{regosi} object}
\usage{
\method{summary}{regosi}(object, ...)
}
\arguments{
\item{object}{A \code{regosi } object}

\item{...}{any default statement}
}
\value{
\code{NULL}
}
\description{
Summary of a \code{regosi } object
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
