% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simpleslopes.R
\name{summary.simpleslopes}
\alias{summary.simpleslopes}
\title{Summary of a \code{simpleslopes} object}
\usage{
\method{summary}{simpleslopes}(slopes)
}
\arguments{
\item{object}{A \code{simpleslopes} object}
}
\value{
\code{NULL}
}
\description{
Summary of a \code{simpleslopes} object
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
