% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_interactions.R
\name{add_interactions}
\alias{add_interactions}
\title{add_interactions}
\usage{
add_interactions(formula, data)
}
\arguments{
\item{formula}{a regression formula including the interaction terms}

\item{data}{a data frame used for the regression}
}
\value{
Object containing the updated regression formula \code{obj$formula}
and data set \code{obj$data}
}
\description{
Creates interaction terms and adds them to the data frame
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
