% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_samples.R
\name{boot_samples}
\alias{boot_samples}
\title{Calculates ordinary bootstrap samples}
\usage{
boot_samples(data, R = 5000)
}
\arguments{
\item{data}{A data.frame used for the regression}

\item{R}{Number of bootstrap samples. Default is 5000.}
}
\value{
List containing R bootstrap samples
}
\description{
This is a function that calculates a given amount of
ordinary bootstrap samples from a data.frame.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
