% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi.R, R/bootmi.lm.R
\name{bootmi}
\alias{bootmi}
\alias{bootmi.lm}
\title{bootmi: Bootstrap multiple imputed survey data}
\usage{
bootmi(formula, data, R, impute = c("none", "norm.predict", "pmm", "mean"),
  center_mods = FALSE, seed = FALSE, parallel = FALSE, resint = FALSE)

\method{bootmi}{lm}(bootmi)
}
\arguments{
\item{formula}{A regression formula}

\item{data}{A data.set}

\item{R}{Number of bootstraps.}

\item{impute}{Deterministic imputation method of type 
"none", "norm.predict", "pmm", "mean"}

\item{center_mods}{TRUE or FALSE}

\item{seed}{Value for set.seed, default = FALSE}

\item{parallel}{TRUE or FALSE}

\item{resint}{TRUE or FALSE}

\item{bootmi}{An object of class bootmi}
}
\description{
ordinary bootstrap for deterministic imputation methods

The method regresses the model on original and bootstrapped data.
For faster computation, the parallel-package is implemented.
}
\details{
The method is an implementation of bootstrapping imputed data  
as proposed by Shao and Sitter (1996). In addition, the method 
makes optional use of residual centering as proposed by 
Little, Bovaird and Widaman (2006), MICE as provided by 
van Buuren and Groothuis-Oudshoorn (2011), and takes into account
the transform, then impute procedure as proposed by von Hippel (2009).
For faster computation, the parallel-package is implemented.
}
\references{
Buuren, Stef van; Groothuis-Oudshoorn, Karin (2011): mice. 
Multivariate Imputation by Chained Equations in R. In: 
Journal of Statistical Software 45 (3). DOI: 10.18637/jss.v045.i03.

Hippel, Paul T. von (2009): How to Impute Interactions, 
Squares, and other Transformed Variables. In: Sociological Methodology 
39 (1), S. 265–291. DOI: 10.1111/j.1467-9531.2009.01215.x.

Little, Todd D.; Bovaird, James A.; Widaman, Keith F. (2006): 
On the merits of orthogonalizing powered and product terms: 
Implications for modeling interactions among latent variables. 
In: Structural Equation Modeling 13 (4), S. 497–519.

R Core Team (2017). R: A language and environment for 
statistical computing. R Foundation for Statistical Computing, 
Vienna, Austria. URL https://www.R-project.org/.

Shao, Jun; Sitter, Randy R. (1996): 
Bootstrap for Imputed Survey Data. In: Journal of the 
American Statistical Association 91 (435), S. 1278–1288. 
DOI: 10.2307/2291746.
}
\author{
Stephan Volpers <stephan.volpers@plixed.de>

Stephan Volpers <stephan.volpers@plixed.de>
}
