% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_residual.R
\name{build_residual}
\alias{build_residual}
\title{calculates values of residual interaction term}
\usage{
build_residual(variable, data, new_varname)
}
\arguments{
\item{variable}{an interaction term}

\item{data}{a data.frame}

\item{new_varname}{name of the residual interaction term}
}
\value{
A data.frame of one row consisting of the residual interaction term.
}
\description{
This function is a helper function and needed to create residual interaction
terms out of the regression model. It takes a data.frame and returns a new 
data.frame with one row in which the mean or mode of the columns is reported.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
