% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simslop.R
\name{numformat}
\alias{numformat}
\title{Summary of a \code{simslop} object}
\usage{
numformat(val, decpoints = 2)
}
\arguments{
\item{object}{A \code{simslop} object}

\item{...}{any default statement}
}
\value{
\code{NULL}
}
\description{
Summary of a \code{simslop} object
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
