% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simslop.R
\name{plot}
\alias{plot}
\alias{plot.simslop}
\alias{plot.simslop.bootmi}
\title{plots simple slopes}
\usage{
\method{plot}{simslop}(x, ...)

\method{plot}{simslop.bootmi}(x, ...)
}
\arguments{
\item{x}{An object of type \code{simslop} or 
\code{simslop.bootmi}}

\item{...}{any default arguments}
}
\value{
\code{NULL}
}
\description{
Opens a new graphical window and plots simple slopes. 
CAUTION: 
No more than 10 different values of the moderator allowed. 
I recommend a maximum of six to plot.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
