% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simpleslopes.R
\name{plot}
\alias{plot}
\alias{plot.simpleslopes}
\alias{plot.simpleslopes.bootmi}
\title{plots simple slopes}
\usage{
\method{plot}{simpleslopes}(x, ...)

\method{plot}{simpleslopes.bootmi}(x, ...)
}
\arguments{
\item{x}{An object of type \code{simpleslopes} or 
\code{simpleslopes.bootmi}}

\item{...}{any default statement}
}
\value{
\code{NULL}
}
\description{
Opens a new graphical window and plots simple slopes
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
