% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simpleslopes.R
\name{plot}
\alias{plot}
\alias{plot.simpleslopes}
\alias{plot.simpleslopes.bootmi}
\title{plots simple slopes}
\usage{
\method{plot}{simpleslopes}(slopes)

\method{plot}{simpleslopes.bootmi}(obj)
}
\arguments{
\item{slopes}{An object of type \code{simpleslopes} or 
\code{simpleslopes.bootmi}}
}
\value{
\code{NULL}
}
\description{
Opens a new graphical window and plots simple slopes
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
