% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootmi.R
\name{print.bootmi}
\alias{print.bootmi}
\alias{print.simpleslopes.bootmi}
\alias{print.bootmi.lm}
\alias{print.regosi}
\alias{print.simpleslopes}
\title{Prints a \code{bootmi} object}
\usage{
\method{print}{bootmi}(x, ...)

\method{print}{simpleslopes.bootmi}(x, ...)

\method{print}{bootmi.lm}(x, ...)

\method{print}{regosi}(x, ...)

\method{print}{simpleslopes}(x, ...)
}
\arguments{
\item{x}{A \code{bootmi} object}

\item{...}{Other print arguments}
}
\value{
\code{NULL}
}
\description{
Prints a \code{bootmi} object

Prints a \code{simpleslopes.bootmi} object

Prints a \code{bootmi.lm} object

Prints a \code{regosi} object

Prints a \code{simpleslopes} object
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
