% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regosi.R
\name{regosi}
\alias{regosi}
\alias{regosi.default}
\alias{regosi.lmerMod}
\alias{regosi.lm}
\alias{regosi.mira}
\title{Calculates regions of significance for moderation effects}
\usage{
regosi(object, x_var, m_var, ci = 95)

\method{regosi}{default}(object, x_var, m_var, ci = 95)

\method{regosi}{lmerMod}(object, x_var, m_var, ci)

\method{regosi}{lm}(object, x_var, m_var, ci)

\method{regosi}{mira}(object, x_var, m_var, ci)
}
\arguments{
\item{object}{Object, if availiable but NOT for default}

\item{x_var}{Name of the independend variable}

\item{m_var}{Name of the moderating variable}

\item{ci}{Confidence interval, default 95}

\item{coeff}{Coefficients from fitted model, just for default}

\item{dat}{Data used to fit model, just for default}

\item{cov_matrix}{Variance-Covariance-Matrix of fitted model, 
just for default}
}
\value{
object of class "regosi" including
\itemize{
  \item $a, Indicator for Interval
  \item $b, Value of slope
  \item $root_term, Indicator for significant regions
  \item $m_var, Name of the moderating variable
  \item $x_var, Name of the independend variable
}
}
\description{
The method is an implementation of regions of significance  
as proposed by Bauer and Curran (2005). The default takes 
parameters needed to calculate regions of significance. For objects 
of class "lm", "lmerMod" and "mira" exist helper functions 
that extract parameters "coeff", "dat", "cov_matrix" from the object.
}
\section{Helper functions exist}{

Uses object of classes lm, lmerMod or mira 
and extracts coefficients, data, and variance covariance matrix
}

\references{
Bauer, Daniel J.; Curran, Patrick J. (2005): 
Probing Interactions in Fixed and Multilevel Regression: 
Inferential and Graphical Techniques. In: Multivariate Behavioral Research
40 (3), S. 373–400. DOI: 10.1207/s15327906mbr4003_5.
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
