% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootmi.lm.R
\name{summary.bootmi.lm}
\alias{summary.bootmi.lm}
\alias{summary.simslop.bootmi}
\title{Summary of a \code{bootmi.lm} object}
\usage{
\method{summary}{bootmi.lm}(object, ci_type = c("basic", "norm", "perc",
  "bca"), ci = 95, ...)

\method{summary}{simslop.bootmi}(object, ci_type = c("basic", "norm", "stud",
  "perc", "bca", "all"), ci = 95, ...)
}
\arguments{
\item{object}{A \code{bootmi.lm} object}

\item{ci_type}{Type of confidence interval, namely 
"basic", "norm", "stud", "perc", "bca", or "all"}

\item{ci}{Confidence interval, default 95}

\item{...}{Other summary arguments}
}
\value{
\code{NULL}
}
\description{
Summary of a \code{bootmi.lm} object

Summary of a \code{simslop.bootmi} object
}
\author{
Stephan Volpers \email{stephan.volpers@plixed.de}
}
